requirements.txt
absl-py==2.1.0
alembic==1.13.1
aniso8601==9.0.1
antlr4-python3-runtime==4.8
arrow==1.3.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
auth0-python==4.3.0
binaryornot==0.4.4
blessed==1.20.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
coloredlogs==14.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.2.1
cookiecutter==2.1.1
cryptography==42.0.3
cycler==0.12.1
debugpy @ file:///croot/debugpy_1690905042057/work
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.6.1
docker==7.0.0
docker-pycreds==0.4.0
email-validator==2.0.0
entrypoints==0.4
exceptiongroup==1.2.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
filelock==3.13.4
Flask==2.2.5
Flask-JWT-Extended==4.4.3
Flask-SQLAlchemy==2.5.1
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
gpustat @ file:///croot/gpustat_1701888826609/work
gpytorch==1.12
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.51.1
gunicorn==22.0.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1708996548741/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1709559745751/work
itsdangerous==2.2.0
jaxtyping==0.2.33
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
jeepney==0.8.0
Jinja2==3.1.3
jinja2-time==0.2.0
joblib==1.4.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1710255804825/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1710257277185/work
keyring==23.4.1
keyrings.alt==3.1
kiwisolver==1.4.5
linear-operator==0.5.2
Mako==1.3.5
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
-e git+https://github.com/aristizabal95/medperf-2.git@d0e4ee18eddc25b42ad36b0ba97c4e1e975b510e#egg=medperf&subdirectory=cli
merge-args==0.1.4
mlcube @ git+https://github.com/hasan7n/mlcube@6b5ede46a66e7ab1c1df29bf57160be83e330fff#subdirectory=mlcube
mlcube_docker @ git+https://github.com/hasan7n/mlcube@6b5ede46a66e7ab1c1df29bf57160be83e330fff#subdirectory=runners/mlcube_docker
mlcube_singularity @ git+https://github.com/hasan7n/mlcube@6b5ede46a66e7ab1c1df29bf57160be83e330fff#subdirectory=runners/mlcube_singularity
mlflow==2.14.1
mpmath==1.3.0
msgpack==1.0.8
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
netcal==1.3.5
networkx==3.3
numpy==1.26.4
nvflare==2.4.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py @ file:///croot/nvidia-ml-py_1701875136273/work
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.1.1
opencv-python==4.10.0.84
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1706713388748/work
pluggy==1.4.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==3.20.3
psutil==5.9.8
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==15.0.2
pycparser==2.21
pydantic==1.10.2
pydicom==2.4.4
pyfakefs==5.0.0
Pygments==2.17.2
pyhocon==0.3.60
PyJWT==2.8.0
pyparsing==3.1.2
pyro-api==0.1.2
pyro-ppl==1.9.1
pytest==8.0.1
pytest-mock==1.13.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
querystring-parser==1.2.4
requests==2.31.0
rich==13.7.0
schema==0.7.5
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
SecretStorage==3.3.3
semver==3.0.2
sentry-sdk==2.6.0
setproctitle==1.3.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
SQLAlchemy==1.4.31
sqlparse==0.5.0
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
stevedore==5.1.0
sympy==1.12
synapseclient==2.7.0
tenseal==0.3.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.5.0
tikzplotlib==0.10.1
time-machine==2.4.0
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1708363098266/work
tqdm==4.66.4
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
triton==2.2.0
typeguard==2.13.3
typer==0.6.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
validators==0.18.2
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
wandb==0.17.3
watchdog==3.0.0
wcwidth==0.2.13
webcolors==24.6.0
websockets==12.0
Werkzeug==2.2.2
wrapt==1.16.0
yaspin==2.1.0
zipp==3.17.0
